backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
passlib==1.7.4
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.2
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
